put files into dirs
