+++ .gitignore
@@ -135,6 +135,7 @@ dmypy.json
 /.pnp
 .pnp.js
 pnpm-lock.yaml
+.pnpm-store
 
 # testing
 /coverage
