+++ client/component/src/InteractionModeToggle.tsx
@@ -3,7 +3,7 @@ import { IoShapesOutline } from 'react-icons/io5';
 import { TbZoomInArea, TbZoomPan } from 'react-icons/tb';
 
 import { ToggleGroup } from '@h5web/lib';
-import type { IIconType } from './Modal';
+import type { IIconType } from './modals/Modal';
 import { InteractionModeType } from './utils';
 
 /**
