+++ client/component/src/modals/AxisConfigModal.tsx
@@ -13,11 +13,11 @@ import {
 import type { TypedArray } from 'ndarray';
 import type { ReactNode } from 'react';
 
-import DomainConfig from './DomainConfig';
-import LabelledInput from './LabelledInput';
+import DomainConfig from '../DomainConfig';
+import LabelledInput from '../LabelledInput';
 import type { IIconType } from './Modal';
 import Modal from './Modal';
-import { createHistogramParams } from './utils';
+import { createHistogramParams } from '../utils';
 
 type EnumArray<T> = Array<T[keyof T]>;
