+++ client/component/src/plots/ConnectedPlot.tsx
@@ -5,7 +5,7 @@ import useWebSocket, { ReadyState } from 'react-use-websocket';
 import { toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
-import AnyPlot from './AnyPlot';
+import AnyPlot from '../AnyPlot';
 import {
   appendDLineData,
   calculateMultiXDomain,
@@ -18,8 +18,8 @@ import {
   createDTableData,
   isHeatmapData,
   measureInteraction,
-} from './utils';
-import { recreateSelection, type SelectionBase } from './selections/utils';
+} from '../utils';
+import { recreateSelection, type SelectionBase } from '../selections/utils';
 import type {
   AnyPlotProps,
   AxesParameters,
@@ -28,13 +28,13 @@ import type {
   DLineData,
   HeatmapPlotProps,
   SurfacePlotProps,
-} from './AnyPlot';
+} from '../AnyPlot';
 import type { LineData } from './LinePlot';
 import type { ImageData } from './ImagePlot';
 import type { ScatterData } from './ScatterPlot';
 import type { SurfaceData } from './SurfacePlot';
-import type { TableData } from './TableDisplay';
-import type { DHeatmapData } from './utils';
+import type { TableData } from '../TableDisplay';
+import type { DHeatmapData } from '../utils';
 
 type MsgType =
   | 'status'
