+++ client/component/src/selection-components/PolygonalSelectionConfig.module.css
