+++ client/component/src/selection-components/PolygonalSelectionConfig.tsx
@@ -1,6 +1,6 @@
-import type PolygonalSelection from './selections/PolygonalSelection';
+import type PolygonalSelection from '../selections/PolygonalSelection';
 import { PointXInput, PointYInput } from './SelectionConfigComponents';
-import type { SelectionBase } from './selections/utils';
+import type { SelectionBase } from '../selections/utils';
 
 import styles from './PolygonalSelectionConfig.module.css';
